Factoring out service api layer for client side - first pass
