+++ src/static/index.html
@@ -53,6 +53,8 @@
                     </li>
                 </ul>
             </div>
+
+            <canvas id="myChart" width="400px" height="400px"></canvas>
         </div>
 
         <div class="content" id="instanceDetails">
